tools/requirements.txt
google-generativeai==0.8.3
pydantic==2.9.2
python-dateutil==2.9.0
